requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
cairocffi==1.7.1
cairosvg==2.8.2
certifi==2025.11.12
cffi==2.0.0
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.1
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-telegram-bot==22.5
sniffio==1.3.1
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
webencodings==0.5.1
python-Levenshtein==0.26.1